push ui components to separate folder
