+++ components/admin-dashboard/admin-dashboard-empty.tsx
@@ -1,6 +1,6 @@
 import { Box, Flex, Image, Text } from "@chakra-ui/react";
 import DiagonalRectangle from "../nft-sale-widget/commons/diagonal-rectangle";
-import PsyButton from "../psy-button";
+import PsyButton from "../ui/psy-button";
 
 const AdminDashboardEmptyState = ({
   setOpenCreateSale
