+++ components/nft-sale-widget/layout/nft-sale/mint-specific-section.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Box, Grid, Image, Text } from "@chakra-ui/react";
 import NFTPrice from "@/components/commons/nftprice";
-import MintButton from "@/components/mint-button";
+import MintButton from "@/components/ui/mint-button";
 
 export const MintSpecificSection = () => {
   return <Box textAlign="center" py={4} px={4}></Box>;
