+++ components/ui/marquee.tsx
@@ -1,7 +1,7 @@
 import { Fragment } from "react";
 import { Box, type BoxProps, keyframes, Text } from "@chakra-ui/react";
 
-import { Logo } from "@/components/icons";
+import { Logo } from "@/components/ui/icons";
 import { useDimensions } from "@/lib/hooks";
 
 const scroll = keyframes`
